+++ CHANGELOG.md
@@ -5,6 +5,41 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-[Tags on this repository](https://github.com/appulse-projects/utils-java/tags)
+[Tags on this repository](https://github.com/xxlabaza/median-meter/tags)
 
 ## [Unreleased]
+
+- Add median function implementation.
+
+## [0.2.0](https://github.com/xxlabaza/median-meter/releases/tag/0.2.0) - 2019-01-15
+
+Use Hazelcast for application clusterization.
+
+### Added
+
+- Custom Hazelcast cluster, which use `MQTT`-based service discovery;
+- Unit and integration tests;
+- JavaDocs.
+
+### Changed
+
+- Enable JavaDoc's checkstyle rules;
+- Maven dependencies.
+
+## [0.1.0](https://github.com/xxlabaza/median-meter/releases/tag/0.1.0) - 2019-01-14
+
+Discovery.
+
+### Added
+
+- Service discovery interface and its `MQTT` implementation;
+- Unit and integration tests;
+- JavaDocs.
+
+### Changed
+
+- Maven dependencies.
+
+## [0.0.1](https://github.com/xxlabaza/median-meter/releases/tag/0.0.1) - 2019-01-14
+
+Initial release